Use msgspec as default serializer (#40)
* feat: User msgspec as default serializer

* chore: make linter happy

* Add stability test with memory profiler

* Update readme

* feat: separate local and dockerize benchmarks

+++ zero/client_server/server.py
@@ -39,7 +39,7 @@ class ZeroServer:
             Port of the ZeroServer.
         encoder: Optional[Encoder]
             Encoder to encode/decode messages from/to client.
-            Default is msgpack.
+            Default is msgspec.
             If any other encoder is used, the client should use the same encoder.
             Implement custom encoder by inheriting from `zero.encoder.Encoder`.
         """
@@ -183,7 +183,11 @@ class ZeroServer:
 
     @util.log_error
     def _remove_ipc(self):
-        if os.name == "posix":
+        if (
+            os.name == "posix"
+            and self._device_ipc is not None
+            and os.path.exists(self._device_ipc)
+        ):
             os.remove(self._device_ipc)
 
     @util.log_error
